chore: rename EagerXml -> ObjectXml
